feat: add Elasticsearch integration for goods indexing and search functionality

+++ oneline-shop-framework/src/main/java/com/onlineshop/framework/common/entity/PageQueryDTO.java
@@ -24,11 +24,11 @@ public class PageQueryDTO {
      */
     @JsonProperty("page")
     @Builder.Default
-    private Long pageNo = 1L;
+    private Integer pageNo = 1;
 
     /**
      * 每页数量
      */
     @Builder.Default
-    private Long pageSize = 10L;
+    private Integer pageSize = 10;
 }
